feat: add toc configuration and collapsible management
