+++ assets/sass/theme/variables.sass
@@ -24,14 +24,6 @@ $helper-loading-color: mediumseagreen
 $modal-max-height: 75%
 $modal-width: 75%
 //
-// Toc variables
-//
-$toc-font-size: 0.875rem
-$toc-padding: 0 0 0 $theme-gap-medium-size
-$toc-width: 300px
-$toc-taxonomy-color: silver
-$toc-taxonomy-color-hover: smooth-color($theme-color, 20)
-//
 // Navbar variables
 //
 $navbar-item-padding: calc((#{$navbar-height} - #{$control-height}) / 2)
@@ -58,6 +50,18 @@ $scrollbar-color: smooth-color($sidebar-background, 20)
 $scrollbar-color-hover: smooth-color($sidebar-background, 30)
 $scrollbar-color-action: smooth-color($sidebar-background, 40)
 $scrollbar-rail-size: 0.5rem
+//
+// Toc variables
+//
+$toc-font-size: 0.875rem
+$toc-padding: 0 0 0 $theme-gap-medium-size
+$toc-width: 275px
+$toc-collapsible-width: 25px
+$toc-taxonomy-color: silver
+$toc-taxonomy-color-hover: smooth-color($theme-color, 20)
+$toc-taxonomy-border-color: smooth-color($content-writing-color, 85)
+$toc-taxonomy-collapsible-color: smooth-color($content-writing-color, 85)
+$toc-taxonomy-collapsible-color-hover: smooth-color($toc-taxonomy-collapsible-color, 10)
 //
 // Shortcodes variables
 //
