+++ exampleSite/templates/configuration/config.fr.toml
@@ -1,17 +1,19 @@
 [params]
   # Chemin des fichiers logo du site
-  logo             = "images/logo.png"
-  logoTouch        = "images/logoTouch.png"
+  logo               = "images/logo.png"
+  logoTouch          = "images/logoTouch.png"
   # Chemin du fichier de l'image affichée en cas d'erreur 404
-  image404         = "images/404.gif"
+  image404           = "images/404.gif"
   # Chemin du fichier du favicon du site
-  favicon          = "images/favicon.png"
+  favicon            = "images/favicon.png"
   # Aperçu par défaut de la page d'accueil
-  homeLayout       = "onboarding"
+  homeLayout         = "onboarding"
   # Ordre du contenu du site (https://gohugo.io/templates/lists/#order-content)
-  siteContentOrder = "weight"
+  siteContentOrder   = "weight"
+  # Active/Désactive la table des matières de manière globale (sur tout le site)
+  toc                = true
   # Active une bannière globale sur toutes les pages du site
-  globalBanner    = true
+  globalBanner       = true
   # Version courante du site
   currentVersion     = "latest"
   # Url de la dernière version du site
