+++ layouts/partials/theme/sidebar.html
@@ -9,14 +9,14 @@
     </div>
     <div id='sidebarCollapsible' class='is-marginless'>
       <div class='card is-uncollapse-action'>
-        <div id='sidebarUncollapse' class='card-header is-single-link'>
+        <div id='sidebarUncollapse' class='card-header is-single-link' title='{{i18n "sidebar_collapsible_title"}}'>
           <a class='card-header-title'>
             <i class='fa-solid fa-angles-right fa-lg'></i>
           </a>
         </div>
       </div>
       <div class='card is-collapse-action'>
-        <div id='sidebarCollapse' class='card-header is-single-link'>
+        <div id='sidebarCollapse' class='card-header is-single-link' title='{{i18n "sidebar_collapsible_title"}}'>
           <a class='card-header-title'>
             <i class='fa-solid fa-angles-left fa-lg'></i>
             <p>{{i18n "sidebar_collapse"}}</p>
