+++ layouts/partials/theme/toc.html
@@ -31,3 +31,6 @@
     </div>
   </div>
 </div>
+<div class='column is-sticky is-hidden-touch is-paddingless' id='tocCollapsible' title='{{i18n "toc_collapsible_title"}}'>
+  <i class='fa-solid fa-angles-right fa-lg'></i>
+</div>
